requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
dj-database-url
python-dotenv
whitenoise
stripe
django-cors-headers
Pillow
django-cors-headers
django-filter
python-dotenv
qrcode
django-redis
djangorestframework-simplejwt
celery
redis
matplotlib
reportlab
openpyxl
boto3
django-storages
django-allauth
dj-rest-auth
django-axes
django-otp
django-two-factor-auth
social-auth-app-django
requests
cryptography
google-auth
google-auth-oauthlib
stripe
gunicorn
whitenoise
dj-database-url